pyproject.toml
[project]
name = "threat-intel-backend"
version = "0.1.0"
description = "Collaborative Threat Intelligence Platform - Backend & AI Integration"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
    "psycopg2-binary>=2.9",   
    "alembic>=1.13.0",
    "redis",
    "slowapi",
    "python-jose[cryptography]",
]

[project.optional-dependencies]
dev = ["ruff>=0.1.0", "pytest>=7.0.0", "pytest-asyncio>=0.21.0", "httpx>=0.24.0"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*"]